Crate swc_ecma_ast

Macros

op

Creates a corresponding operator. This macro is used to make code more readable.

Structs

ArrayLit

Array literal.

ArrayPat

ArrowExpr

AssignExpr

AssignPat

AssignPatProp

{key} or {key = value}

AssignProp

AwaitExpr

BigInt

BinExpr

BindingIdent

Identifier used as a pattern.

BlockStmt

Use when only block statements are allowed.

Bool

A boolean literal.

BreakStmt

CallExpr

CatchClause

Class

ClassDecl

ClassExpr

Class expression.

ClassMethod

ClassProp

ComputedPropName

CondExpr

Constructor

ContinueStmt

DebuggerStmt

Decorator

DoWhileStmt

EmptyStmt

ExportAll

export * from 'mod'

ExportDecl

ExportDefaultDecl

ExportDefaultExpr

ExportDefaultSpecifier

ExportNamedSpecifier

ExportNamespaceSpecifier

export * as foo from 'src';

ExprOrSpread

ExprStmt

FnDecl

FnExpr

Function expression.

ForInStmt

ForOfStmt

ForStmt

Function

Common parts of function and method.

GetterProp

Ident

A complete identifier with span.

IfStmt

Import

ImportDecl

ImportDefaultSpecifier

e.g. import foo from 'mod.js'

ImportNamedSpecifier

e.g. local = foo, imported = None import { foo } from 'mod.js' e.g. local = bar, imported = Some(foo) for import { foo as bar } from 'mod.js'

ImportStarAsSpecifier

e.g. import * as foo from 'mod.js'.

Invalid

Represents a invalid node.

JSXAttr

JSXClosingElement

JSXClosingFragment

JSXElement

JSXEmptyExpr

JSXExprContainer

JSXFragment

JSXMemberExpr

JSXNamespacedName

XML-based namespace syntax:

JSXOpeningElement

JSXOpeningFragment

JSXSpreadChild

JSXText

KeyValuePatProp

{key: value}

KeyValueProp

LabeledStmt

ListFormat

Represents the formatting rule for a list of nodes.

MemberExpr

MetaPropExpr

MethodProp

Module

NamedExport

export { foo } from 'mod' export { foo as bar } from 'mod'

NewExpr

Null

Number

A numeric literal.

ObjectLit

Object literal.

ObjectPat

OptCall

OptChainExpr

Param

ParenExpr

PrivateMethod

PrivateName

PrivateProp

Regex

ReservedUnused

RestPat

EsTree RestElement

ReturnStmt

Script

SeqExpr

SetterProp

SpreadElement

StaticBlock

Str

A string literal.

Super

SuperPropExpr

SwitchCase

SwitchStmt

TaggedTpl

ThisExpr

ThrowStmt

Tpl

TplElement

TryStmt

TsArrayType

TsAsExpr

TsCallSignatureDecl

TsConditionalType

TsConstAssertion

TsConstructSignatureDecl

TsConstructorType

TsEnumDecl

TsEnumMember

TsExportAssignment

TypeScript’s own parser uses ExportAssignment for both export default and export =. But for @babel/parser, export default is an ExportDefaultDecl, so a TsExportAssignment is always export =.

TsExprWithTypeArgs

TsExternalModuleRef

TsFnType

TsGetterSignature

TsImportEqualsDecl

TsImportType

TsIndexSignature

TsIndexedAccessType

TsInferType

TsInstantiation

TsInterfaceBody

TsInterfaceDecl

TsIntersectionType

TsKeywordType

TsLitType

TsMappedType

TsMethodSignature

TsModuleBlock

TsModuleDecl

TsNamespaceDecl

TsNamespaceExportDecl

TsNonNullExpr

TsOptionalType

TsParamProp

TsParenthesizedType

TsPropertySignature

TsQualifiedName

TsRestType

TsSatisfiesExpr

TsSetterSignature

TsThisType

TsTplLitType

TsTupleElement

TsTupleType

TsTypeAliasDecl

TsTypeAnn

TsTypeAssertion

TsTypeLit

TsTypeOperator

TsTypeParam

TsTypeParamDecl

TsTypeParamInstantiation

TsTypePredicate

TsTypeQuery

typeof operator

TsTypeRef

TsUnionType

UnaryExpr

UpdateExpr

VarDecl

VarDeclarator

WhileStmt

WithStmt

YieldExpr

Enums

Accessibility

AssignOp

BinaryOp

BlockStmtOrExpr

Callee

ClassMember

Decl

DefaultDecl

EsVersion

ExportSpecifier

Expr

ImportSpecifier

JSXAttrName

JSXAttrOrSpread

JSXAttrValue

JSXElementChild

JSXElementName

JSXExpr

JSXObject

Used for obj property of JSXMemberExpr.

Lit

MemberProp

MetaPropKind

MethodKind

ModuleDecl

ModuleExportName

ModuleItem

ObjectPatProp

OptChainBase

ParamOrTsParamProp

Pat

PatOrExpr

Program

Prop

PropName

PropOrSpread

Stmt

SuperProp

TruePlusMinus

TsEntityName

TsEnumMemberId

Invalid: Ident with empty symbol.

TsFnOrConstructorType

TsFnParam

TsKeywordTypeKind

TsLit

TsModuleName

TsModuleRef

TsNamespaceBody

namespace A.B { } is a namespace named A with another TsNamespaceDecl as its body.

TsParamPropParam

TsThisTypeOrIdent

TsType

TsTypeElement

TsTypeOperatorOp

TsTypeQueryExpr

TsUnionOrIntersectionType

UnaryOp

UpdateOp

VarDeclKind

VarDeclOrExpr

VarDeclOrPat

Traits

IdentExt

SourceMapperExt

SpanExt

Type Aliases

Id

See Ident for documentation.