Enum swc_ecma_ast::Decl

pub enum Decl {
    Class(ClassDecl),
    Fn(FnDecl),
    Var(VarDecl),
    TsInterface(TsInterfaceDecl),
    TsTypeAlias(TsTypeAliasDecl),
    TsEnum(TsEnumDecl),
    TsModule(TsModuleDecl),
}

Variants

Class(ClassDecl)

Fn(FnDecl)

Var(VarDecl)

TsInterface(TsInterfaceDecl)

TsTypeAlias(TsTypeAliasDecl)

TsEnum(TsEnumDecl)

TsModule(TsModuleDecl)

Implementations

impl Decl

pub fn is_class(&self) -> bool

Returns true if self is of variant Class.

pub fn expect_class(self) -> ClassDecl where
Self: Debug,

Unwraps the value, yielding the content of Class.

Panics

Panics if the value is not Class, with a panic message including the content of self.

pub fn class(self) -> Option<ClassDecl>

Returns Some if self is of variant Class, and None otherwise.

pub fn is_fn_decl(&self) -> bool

Returns true if self is of variant Fn.

pub fn expect_fn_decl(self) -> FnDecl where
Self: Debug,

Unwraps the value, yielding the content of Fn.

Panics

Panics if the value is not Fn, with a panic message including the content of self.

pub fn fn_decl(self) -> Option<FnDecl>

Returns Some if self is of variant Fn, and None otherwise.

pub fn is_var(&self) -> bool

Returns true if self is of variant Var.

pub fn expect_var(self) -> VarDecl where
Self: Debug,

Unwraps the value, yielding the content of Var.

Panics

Panics if the value is not Var, with a panic message including the content of self.

pub fn var(self) -> Option<VarDecl>

Returns Some if self is of variant Var, and None otherwise.

pub fn is_ts_interface(&self) -> bool

Returns true if self is of variant TsInterface.

pub fn expect_ts_interface(self) -> TsInterfaceDecl where
Self: Debug,

Unwraps the value, yielding the content of TsInterface.

Panics

Panics if the value is not TsInterface, with a panic message including the content of self.

pub fn ts_interface(self) -> Option<TsInterfaceDecl>

Returns Some if self is of variant TsInterface, and None otherwise.

pub fn is_ts_type_alias(&self) -> bool

Returns true if self is of variant TsTypeAlias.

pub fn expect_ts_type_alias(self) -> TsTypeAliasDecl where
Self: Debug,

Unwraps the value, yielding the content of TsTypeAlias.

Panics

Panics if the value is not TsTypeAlias, with a panic message including the content of self.

pub fn ts_type_alias(self) -> Option<TsTypeAliasDecl>

Returns Some if self is of variant TsTypeAlias, and None otherwise.

pub fn is_ts_enum(&self) -> bool

Returns true if self is of variant TsEnum.

pub fn expect_ts_enum(self) -> TsEnumDecl where
Self: Debug,

Unwraps the value, yielding the content of TsEnum.

Panics

Panics if the value is not TsEnum, with a panic message including the content of self.

pub fn ts_enum(self) -> Option<TsEnumDecl>

Returns Some if self is of variant TsEnum, and None otherwise.

pub fn is_ts_module(&self) -> bool

Returns true if self is of variant TsModule.

pub fn expect_ts_module(self) -> TsModuleDecl where
Self: Debug,

Unwraps the value, yielding the content of TsModule.

Panics

Panics if the value is not TsModule, with a panic message including the content of self.

pub fn ts_module(self) -> Option<TsModuleDecl>

Returns Some if self is of variant TsModule, and None otherwise.