Enum swc_ecma_ast::ExprKind

pub enum ExprKind {
    This,
    Array(ArrayLit),
    Object(ObjectLit),
    Fn(FnExpr),
    Unary(UnaryExpr),
    Update(UpdateExpr),
    Bin(BinExpr),
    Assign(AssignExpr),
    Member(MemberExpr),
    Cond(CondExpr),
    Call(CallExpr),
    New(NewExpr),
    Seq(SeqExpr),
    Ident(Ident),
    Lit(Lit),
    Tpl(TplLit),
    Arrow(ArrowExpr),
    Class(ClassExpr),
    Yield(YieldExpr),
    MetaProp(MetaPropExpr),
    Await(AwaitExpr),
    Paren(Box<Expr>),
}

Variants
ThisArray(ArrayLit)Object(ObjectLit)Fn(FnExpr)Unary(UnaryExpr)Update(UpdateExpr)

++v, --v, v++, v--

Bin(BinExpr)Assign(AssignExpr)Member(MemberExpr)

A member expression. If computed is true, the node corresponds to a computed (a[b]) member expression and property is an Expression. If computed is false, the node corresponds to a static (a.b) member expression and property is an Identifier.

Cond(CondExpr)

true ? 'a' : 'b'

Call(CallExpr)New(NewExpr)

new Cat()

Seq(SeqExpr)Ident(Ident)Lit(Lit)Tpl(TplLit)Arrow(ArrowExpr)Class(ClassExpr)Yield(YieldExpr)MetaProp(MetaPropExpr)Await(AwaitExpr)Paren(Box<Expr>)