Enum swc_ecma_ast::Pat

pub enum Pat {
    Ident(BindingIdent),
    Array(ArrayPat),
    Rest(RestPat),
    Object(ObjectPat),
    Assign(AssignPat),
    Invalid(Invalid),
    Expr(Box<Expr>),
}

Variants

Ident(BindingIdent)

Tuple Fields

0: BindingIdent

Array(ArrayPat)

Tuple Fields

0: ArrayPat

Rest(RestPat)

Tuple Fields

0: RestPat

Object(ObjectPat)

Tuple Fields

0: ObjectPat

Assign(AssignPat)

Tuple Fields

0: AssignPat

Invalid(Invalid)

Tuple Fields

0: Invalid

Expr(Box<Expr>)

Tuple Fields

0: Box<Expr>

Only for for-in / for-of loops. This is syntactically valid.

Implementations

impl Pat

pub fn is_ident(&self) -> bool

Returns true if self is of variant Ident.

pub fn expect_ident(self) -> BindingIdent where
Self: Debug,

Unwraps the value, yielding the content of Ident.

Panics

Panics if the value is not Ident, with a panic message including the content of self.

pub fn ident(self) -> Option<BindingIdent>

Returns Some if self is of variant Ident, and None otherwise.

pub fn is_array(&self) -> bool

Returns true if self is of variant Array.

pub fn expect_array(self) -> ArrayPat where
Self: Debug,

Unwraps the value, yielding the content of Array.

Panics

Panics if the value is not Array, with a panic message including the content of self.

pub fn array(self) -> Option<ArrayPat>

Returns Some if self is of variant Array, and None otherwise.

pub fn is_rest(&self) -> bool

Returns true if self is of variant Rest.

pub fn expect_rest(self) -> RestPat where
Self: Debug,

Unwraps the value, yielding the content of Rest.

Panics

Panics if the value is not Rest, with a panic message including the content of self.

pub fn rest(self) -> Option<RestPat>

Returns Some if self is of variant Rest, and None otherwise.

pub fn is_object(&self) -> bool

Returns true if self is of variant Object.

pub fn expect_object(self) -> ObjectPat where
Self: Debug,

Unwraps the value, yielding the content of Object.

Panics

Panics if the value is not Object, with a panic message including the content of self.

pub fn object(self) -> Option<ObjectPat>

Returns Some if self is of variant Object, and None otherwise.

pub fn is_assign(&self) -> bool

Returns true if self is of variant Assign.

pub fn expect_assign(self) -> AssignPat where
Self: Debug,

Unwraps the value, yielding the content of Assign.

Panics

Panics if the value is not Assign, with a panic message including the content of self.

pub fn assign(self) -> Option<AssignPat>

Returns Some if self is of variant Assign, and None otherwise.

pub fn is_invalid(&self) -> bool

Returns true if self is of variant Invalid.

pub fn expect_invalid(self) -> Invalid where
Self: Debug,

Unwraps the value, yielding the content of Invalid.

Panics

Panics if the value is not Invalid, with a panic message including the content of self.

pub fn invalid(self) -> Option<Invalid>

Returns Some if self is of variant Invalid, and None otherwise.

pub fn is_expr(&self) -> bool

Returns true if self is of variant Expr.

pub fn expect_expr(self) -> Box<Expr> where
Self: Debug,

Unwraps the value, yielding the content of Expr.

Panics

Panics if the value is not Expr, with a panic message including the content of self.

pub fn expr(self) -> Option<Box<Expr>>

Returns Some if self is of variant Expr, and None otherwise.