swc_ecma_ast
Enum PatOrExpr

pub enum PatOrExpr {
    Expr(Box<Expr>),
    Pat(Box<Pat>),
}

Variants

Expr(Box<Expr>)

Pat(Box<Pat>)

Implementations

impl PatOrExpr

pub fn pat(self) -> Option<Box<Pat>>

Returns the Pat if this is a pattern, otherwise returns None.

pub fn expr(self) -> Option<Box<Expr>>

Returns the Expr if this is an expression, otherwise returns None.

pub fn expect_pat(self) -> Box<Pat>

pub fn expect_expr(self) -> Box<Expr>

pub fn as_pat(&self) -> Option<&Pat>

pub fn as_expr(&self) -> Option<&Expr>

pub fn is_pat(&self) -> bool

pub fn is_expr(&self) -> bool

pub fn as_ident(&self) -> Option<&Ident>

pub fn as_ident_mut(&mut self) -> Option<&mut Ident>

pub fn normalize_expr(self) -> Self

pub fn normalize_ident(self) -> Self