Enum swc_ecma_ast::PropName

pub enum PropName {
    Ident(Ident),
    Str(Str),
    Num(Number),
    Computed(ComputedPropName),
    BigInt(BigInt),
}

Variants

Ident(Ident)

Tuple Fields

0: Ident

Str(Str)

Tuple Fields

0: Str

String literal.

Num(Number)

Tuple Fields

0: Number

Numeric literal.

Computed(ComputedPropName)

Tuple Fields

0: ComputedPropName

BigInt(BigInt)

Tuple Fields

0: BigInt

Implementations

impl PropName

pub fn is_ident(&self) -> bool

Returns true if self is of variant Ident.

pub fn expect_ident(self) -> Ident where
Self: Debug,

Unwraps the value, yielding the content of Ident.

Panics

Panics if the value is not Ident, with a panic message including the content of self.

pub fn ident(self) -> Option<Ident>

Returns Some if self is of variant Ident, and None otherwise.

pub fn is_str(&self) -> bool

Returns true if self is of variant Str.

pub fn expect_str(self) -> Str where
Self: Debug,

Unwraps the value, yielding the content of Str.

Panics

Panics if the value is not Str, with a panic message including the content of self.

pub fn str(self) -> Option<Str>

Returns Some if self is of variant Str, and None otherwise.

pub fn is_num(&self) -> bool

Returns true if self is of variant Num.

pub fn expect_num(self) -> Number where
Self: Debug,

Unwraps the value, yielding the content of Num.

Panics

Panics if the value is not Num, with a panic message including the content of self.

pub fn num(self) -> Option<Number>

Returns Some if self is of variant Num, and None otherwise.

pub fn is_computed(&self) -> bool

Returns true if self is of variant Computed.

pub fn expect_computed(self) -> ComputedPropName where
Self: Debug,

Unwraps the value, yielding the content of Computed.

Panics

Panics if the value is not Computed, with a panic message including the content of self.

pub fn computed(self) -> Option<ComputedPropName>

Returns Some if self is of variant Computed, and None otherwise.

pub fn is_big_int(&self) -> bool

Returns true if self is of variant BigInt.

pub fn expect_big_int(self) -> BigInt where
Self: Debug,

Unwraps the value, yielding the content of BigInt.

Panics

Panics if the value is not BigInt, with a panic message including the content of self.

pub fn big_int(self) -> Option<BigInt>

Returns Some if self is of variant BigInt, and None otherwise.