swc_ecma_ast

Enum SimpleAssignTarget

#[repr(u32)]
pub enum SimpleAssignTarget {
    Ident(BindingIdent),
    Member(MemberExpr),
    SuperProp(SuperPropExpr),
    Paren(ParenExpr),
    OptChain(OptChainExpr),
    TsAs(TsAsExpr),
    TsSatisfies(TsSatisfiesExpr),
    TsNonNull(TsNonNullExpr),
    TsTypeAssertion(TsTypeAssertion),
    TsInstantiation(TsInstantiation),
    Invalid(Invalid),
}

Variants

Ident(BindingIdent)

Note: This type is to help implementing visitor and the field type_ann is always None.

Member(MemberExpr)

SuperProp(SuperPropExpr)

Paren(ParenExpr)

OptChain(OptChainExpr)

TsAs(TsAsExpr)

TsSatisfies(TsSatisfiesExpr)

TsNonNull(TsNonNullExpr)

TsTypeAssertion(TsTypeAssertion)

TsInstantiation(TsInstantiation)

Invalid(Invalid)

Implementations

impl SimpleAssignTarget

pub const fn is_ident(&self) -> bool

Returns true if self is of variant Ident.

pub fn as_ident(&self) -> Option<&BindingIdent>

Returns Some if self is a reference of variant Ident, and None otherwise.

pub fn as_mut_ident(&mut self) -> Option<&mut BindingIdent>

Returns Some if self is a mutable reference of variant Ident, and None otherwise.

pub fn expect_ident(self) -> BindingIdent
where Self: Debug,

Unwraps the value, yielding the content of Ident.

Panics

Panics if the value is not Ident, with a panic message including the content of self.

pub fn ident(self) -> Option<BindingIdent>

Returns Some if self is of variant Ident, and None otherwise.

pub const fn is_member(&self) -> bool

Returns true if self is of variant Member.

pub fn as_member(&self) -> Option<&MemberExpr>

Returns Some if self is a reference of variant Member, and None otherwise.

pub fn as_mut_member(&mut self) -> Option<&mut MemberExpr>

Returns Some if self is a mutable reference of variant Member, and None otherwise.

pub fn expect_member(self) -> MemberExpr
where Self: Debug,

Unwraps the value, yielding the content of Member.

Panics

Panics if the value is not Member, with a panic message including the content of self.

pub fn member(self) -> Option<MemberExpr>

Returns Some if self is of variant Member, and None otherwise.

pub const fn is_super_prop(&self) -> bool

Returns true if self is of variant SuperProp.

pub fn as_super_prop(&self) -> Option<&SuperPropExpr>

Returns Some if self is a reference of variant SuperProp, and None otherwise.

pub fn as_mut_super_prop(&mut self) -> Option<&mut SuperPropExpr>

Returns Some if self is a mutable reference of variant SuperProp, and None otherwise.

pub fn expect_super_prop(self) -> SuperPropExpr
where Self: Debug,

Unwraps the value, yielding the content of SuperProp.

Panics

Panics if the value is not SuperProp, with a panic message including the content of self.

pub fn super_prop(self) -> Option<SuperPropExpr>

Returns Some if self is of variant SuperProp, and None otherwise.

pub const fn is_paren(&self) -> bool

Returns true if self is of variant Paren.

pub fn as_paren(&self) -> Option<&ParenExpr>

Returns Some if self is a reference of variant Paren, and None otherwise.

pub fn as_mut_paren(&mut self) -> Option<&mut ParenExpr>

Returns Some if self is a mutable reference of variant Paren, and None otherwise.

pub fn expect_paren(self) -> ParenExpr
where Self: Debug,

Unwraps the value, yielding the content of Paren.

Panics

Panics if the value is not Paren, with a panic message including the content of self.

pub fn paren(self) -> Option<ParenExpr>

Returns Some if self is of variant Paren, and None otherwise.

pub const fn is_opt_chain(&self) -> bool

Returns true if self is of variant OptChain.

pub fn as_opt_chain(&self) -> Option<&OptChainExpr>

Returns Some if self is a reference of variant OptChain, and None otherwise.

pub fn as_mut_opt_chain(&mut self) -> Option<&mut OptChainExpr>

Returns Some if self is a mutable reference of variant OptChain, and None otherwise.

pub fn expect_opt_chain(self) -> OptChainExpr
where Self: Debug,

Unwraps the value, yielding the content of OptChain.

Panics

Panics if the value is not OptChain, with a panic message including the content of self.

pub fn opt_chain(self) -> Option<OptChainExpr>

Returns Some if self is of variant OptChain, and None otherwise.

pub const fn is_ts_as(&self) -> bool

Returns true if self is of variant TsAs.

pub fn as_ts_as(&self) -> Option<&TsAsExpr>

Returns Some if self is a reference of variant TsAs, and None otherwise.

pub fn as_mut_ts_as(&mut self) -> Option<&mut TsAsExpr>

Returns Some if self is a mutable reference of variant TsAs, and None otherwise.

pub fn expect_ts_as(self) -> TsAsExpr
where Self: Debug,

Unwraps the value, yielding the content of TsAs.

Panics

Panics if the value is not TsAs, with a panic message including the content of self.

pub fn ts_as(self) -> Option<TsAsExpr>

Returns Some if self is of variant TsAs, and None otherwise.

pub const fn is_ts_satisfies(&self) -> bool

Returns true if self is of variant TsSatisfies.

pub fn as_ts_satisfies(&self) -> Option<&TsSatisfiesExpr>

Returns Some if self is a reference of variant TsSatisfies, and None otherwise.

pub fn as_mut_ts_satisfies(&mut self) -> Option<&mut TsSatisfiesExpr>

Returns Some if self is a mutable reference of variant TsSatisfies, and None otherwise.

pub fn expect_ts_satisfies(self) -> TsSatisfiesExpr
where Self: Debug,

Unwraps the value, yielding the content of TsSatisfies.

Panics

Panics if the value is not TsSatisfies, with a panic message including the content of self.

pub fn ts_satisfies(self) -> Option<TsSatisfiesExpr>

Returns Some if self is of variant TsSatisfies, and None otherwise.

pub const fn is_ts_non_null(&self) -> bool

Returns true if self is of variant TsNonNull.

pub fn as_ts_non_null(&self) -> Option<&TsNonNullExpr>

Returns Some if self is a reference of variant TsNonNull, and None otherwise.

pub fn as_mut_ts_non_null(&mut self) -> Option<&mut TsNonNullExpr>

Returns Some if self is a mutable reference of variant TsNonNull, and None otherwise.

pub fn expect_ts_non_null(self) -> TsNonNullExpr
where Self: Debug,

Unwraps the value, yielding the content of TsNonNull.

Panics

Panics if the value is not TsNonNull, with a panic message including the content of self.

pub fn ts_non_null(self) -> Option<TsNonNullExpr>

Returns Some if self is of variant TsNonNull, and None otherwise.

pub const fn is_ts_type_assertion(&self) -> bool

Returns true if self is of variant TsTypeAssertion.

pub fn as_ts_type_assertion(&self) -> Option<&TsTypeAssertion>

Returns Some if self is a reference of variant TsTypeAssertion, and None otherwise.

pub fn as_mut_ts_type_assertion(&mut self) -> Option<&mut TsTypeAssertion>

Returns Some if self is a mutable reference of variant TsTypeAssertion, and None otherwise.

pub fn expect_ts_type_assertion(self) -> TsTypeAssertion
where Self: Debug,

Unwraps the value, yielding the content of TsTypeAssertion.

Panics

Panics if the value is not TsTypeAssertion, with a panic message including the content of self.

pub fn ts_type_assertion(self) -> Option<TsTypeAssertion>

Returns Some if self is of variant TsTypeAssertion, and None otherwise.

pub const fn is_ts_instantiation(&self) -> bool

Returns true if self is of variant TsInstantiation.

pub fn as_ts_instantiation(&self) -> Option<&TsInstantiation>

Returns Some if self is a reference of variant TsInstantiation, and None otherwise.

pub fn as_mut_ts_instantiation(&mut self) -> Option<&mut TsInstantiation>

Returns Some if self is a mutable reference of variant TsInstantiation, and None otherwise.

pub fn expect_ts_instantiation(self) -> TsInstantiation
where Self: Debug,

Unwraps the value, yielding the content of TsInstantiation.

Panics

Panics if the value is not TsInstantiation, with a panic message including the content of self.

pub fn ts_instantiation(self) -> Option<TsInstantiation>

Returns Some if self is of variant TsInstantiation, and None otherwise.

pub const fn is_invalid(&self) -> bool

Returns true if self is of variant Invalid.

pub fn as_invalid(&self) -> Option<&Invalid>

Returns Some if self is a reference of variant Invalid, and None otherwise.

pub fn as_mut_invalid(&mut self) -> Option<&mut Invalid>

Returns Some if self is a mutable reference of variant Invalid, and None otherwise.

pub fn expect_invalid(self) -> Invalid
where Self: Debug,

Unwraps the value, yielding the content of Invalid.

Panics

Panics if the value is not Invalid, with a panic message including the content of self.

pub fn invalid(self) -> Option<Invalid>

Returns Some if self is of variant Invalid, and None otherwise.

impl SimpleAssignTarget

pub fn leftmost(&self) -> Option<&Ident>