Enum swc_ecma_ast::TsTypeElement

pub enum TsTypeElement {
    TsCallSignatureDecl(TsCallSignatureDecl),
    TsConstructSignatureDecl(TsConstructSignatureDecl),
    TsPropertySignature(TsPropertySignature),
    TsGetterSignature(TsGetterSignature),
    TsSetterSignature(TsSetterSignature),
    TsMethodSignature(TsMethodSignature),
    TsIndexSignature(TsIndexSignature),
}

Variants

TsCallSignatureDecl(TsCallSignatureDecl)

Tuple Fields of TsCallSignatureDecl

0: TsCallSignatureDecl

TsConstructSignatureDecl(TsConstructSignatureDecl)

Tuple Fields of TsConstructSignatureDecl

0: TsConstructSignatureDecl

TsPropertySignature(TsPropertySignature)

Tuple Fields of TsPropertySignature

0: TsPropertySignature

TsGetterSignature(TsGetterSignature)

Tuple Fields of TsGetterSignature

0: TsGetterSignature

TsSetterSignature(TsSetterSignature)

Tuple Fields of TsSetterSignature

0: TsSetterSignature

TsMethodSignature(TsMethodSignature)

Tuple Fields of TsMethodSignature

0: TsMethodSignature

TsIndexSignature(TsIndexSignature)

Tuple Fields of TsIndexSignature

0: TsIndexSignature

Implementations

impl TsTypeElement

pub fn is_ts_call_signature_decl(&self) -> bool

Returns true if self is of variant TsCallSignatureDecl.

pub fn expect_ts_call_signature_decl(self) -> TsCallSignatureDecl where
Self: Debug,

Unwraps the value, yielding the content of TsCallSignatureDecl.

Panics

Panics if the value is not TsCallSignatureDecl, with a panic message including the content of self.

pub fn ts_call_signature_decl(self) -> Option<TsCallSignatureDecl>

Returns Some if self is of variant TsCallSignatureDecl, and None otherwise.

pub fn is_ts_construct_signature_decl(&self) -> bool

Returns true if self is of variant TsConstructSignatureDecl.

pub fn expect_ts_construct_signature_decl(self) -> TsConstructSignatureDecl where
Self: Debug,

Unwraps the value, yielding the content of TsConstructSignatureDecl.

Panics

Panics if the value is not TsConstructSignatureDecl, with a panic message including the content of self.

pub fn ts_construct_signature_decl(self) -> Option<TsConstructSignatureDecl>

Returns Some if self is of variant TsConstructSignatureDecl, and None otherwise.

pub fn is_ts_property_signature(&self) -> bool

Returns true if self is of variant TsPropertySignature.

pub fn expect_ts_property_signature(self) -> TsPropertySignature where
Self: Debug,

Unwraps the value, yielding the content of TsPropertySignature.

Panics

Panics if the value is not TsPropertySignature, with a panic message including the content of self.

pub fn ts_property_signature(self) -> Option<TsPropertySignature>

Returns Some if self is of variant TsPropertySignature, and None otherwise.

pub fn is_ts_getter_signature(&self) -> bool

Returns true if self is of variant TsGetterSignature.

pub fn expect_ts_getter_signature(self) -> TsGetterSignature where
Self: Debug,

Unwraps the value, yielding the content of TsGetterSignature.

Panics

Panics if the value is not TsGetterSignature, with a panic message including the content of self.

pub fn ts_getter_signature(self) -> Option<TsGetterSignature>

Returns Some if self is of variant TsGetterSignature, and None otherwise.

pub fn is_ts_setter_signature(&self) -> bool

Returns true if self is of variant TsSetterSignature.

pub fn expect_ts_setter_signature(self) -> TsSetterSignature where
Self: Debug,

Unwraps the value, yielding the content of TsSetterSignature.

Panics

Panics if the value is not TsSetterSignature, with a panic message including the content of self.

pub fn ts_setter_signature(self) -> Option<TsSetterSignature>

Returns Some if self is of variant TsSetterSignature, and None otherwise.

pub fn is_ts_method_signature(&self) -> bool

Returns true if self is of variant TsMethodSignature.

pub fn expect_ts_method_signature(self) -> TsMethodSignature where
Self: Debug,

Unwraps the value, yielding the content of TsMethodSignature.

Panics

Panics if the value is not TsMethodSignature, with a panic message including the content of self.

pub fn ts_method_signature(self) -> Option<TsMethodSignature>

Returns Some if self is of variant TsMethodSignature, and None otherwise.

pub fn is_ts_index_signature(&self) -> bool

Returns true if self is of variant TsIndexSignature.

pub fn expect_ts_index_signature(self) -> TsIndexSignature where
Self: Debug,

Unwraps the value, yielding the content of TsIndexSignature.

Panics

Panics if the value is not TsIndexSignature, with a panic message including the content of self.

pub fn ts_index_signature(self) -> Option<TsIndexSignature>

Returns Some if self is of variant TsIndexSignature, and None otherwise.