swc_ecma_ast
Struct Ident

pub struct Ident {
    pub span: Span,
    pub sym: JsWord,
    pub optional: bool,
}

A complete identifier with span.

Identifier of swc consists of two parts. The first one is symbol, which is stored using an interned string, JsWord . The second one is SyntaxContext, which can be used to distinguish identifier with same symbol.

Let me explain this with an example.

let a = 5
{
    let a = 3;
}

In the code above, there are two variables with the symbol a.

Other compilers typically uses type like Scope, and store them nested, but in rust, type like Scope requires [Arc<Mutex>] so swc uses different approach. Instead of passing scopes, swc annotates two variables with different tag, which is named SyntaxContext. The notation for the syntax context is #n where n is a number. e.g. foo#1

For the example above, after applying resolver pass, it becomes.

let a#1 = 5
{
    let a#2 = 3;
}

Thanks to the tag we attached, we can now distinguish them.

(JsWord, SyntaxContext)

See Id, which is a type alias for this.

This can be used to store all variables in a module to single hash map.

Comparison

While comparing two identifiers, you can use .to_id().

HashMap

There’s a type named Id which only contains minimal information to distinguish identifiers.

Fields
span: Spansym: JsWordoptional: bool

TypeScript only. Used in case of an optional parameter.

Implementations

impl Ident

pub fn within_ignored_ctxt<F, Ret>(op: F) -> Ret
where F: FnOnce() -> Ret,

In op, EqIgnoreSpan of Ident will ignore the syntax context.

pub fn without_loc(self) -> Ident

Preserve syntax context while drop span.lo and span.hi.

pub fn to_id(&self) -> Id

Creates Id using JsWord and SyntaxContext of self.

pub fn is_valid_start(c: char) -> bool

Returns true if c is a valid character for an identifier start.

pub fn is_valid_continue(c: char) -> bool

Returns true if c is a valid character for an identifier part after start.

pub fn verify_symbol(s: &str) -> Result<(), String>

Alternative for toIdentifier of babel.

Returns Ok if it’s a valid identifier and Err if it’s not valid. The returned Err contains the valid symbol.

pub fn is_dummy(&self) -> bool

impl Ident

pub const fn new(sym: JsWord, span: Span) -> Self