Struct swc_ecma_ast::Number

pub struct Number {
    pub span: Span,
    pub value: f64,
    pub raw: Option<Atom>,
}

A numeric literal.

Creation

If you are creating a numeric literal with a dummy span, please use literal.into(), instead of creating this struct directly.

All of Box<Expr>, Expr, Lit, Number implements From<64> and From<usize>.

Fields
span: Spanvalue: f64

Note: This should not be NaN. Use crate::Ident to represent NaN.

If you store NaN in this field, a hash map will behave strangely.

raw: Option<Atom>

Use None value only for transformations to avoid recalculate characters in number literal