swc_ecma_ast

Type Alias UnsafeId

pub type UnsafeId = (UnsafeAtom, SyntaxContext);

UnsafeId is a wrapper around Id that does not allocate, but extremely unsafe.

Do not use this unless you know what you are doing.

Currently, it’s considered as a unstable API and may be changed in the future without a semver bump.