Enum swc_ecma_ast::ExprKind

pub enum ExprKind {
    This,
    Array {
        elems: Vec<Option<ExprOrSpread>>,
    },
    Object {
        props: Vec<Prop>,
    },
    Function(FnExpr),
    Unary {
        op: UnaryOp,
        prefix: bool,
        arg: Box<Expr>,
    },
    Update {
        op: UpdateOp,
        prefix: bool,
        arg: Box<Expr>,
    },
    Binary {
        op: BinaryOp,
        left: Box<Expr>,
        right: Box<Expr>,
    },
    Assign {
        op: AssignOp,
        left: PatOrExpr,
        right: Box<Expr>,
    },
    Member {
        obj: ExprOrSuper,
        prop: Box<Expr>,
        computed: bool,
    },
    Cond {
        test: Box<Expr>,
        cons: Box<Expr>,
        alt: Box<Expr>,
    },
    Call {
        callee: ExprOrSuper,
        args: Vec<ExprOrSpread>,
    },
    New {
        callee: Box<Expr>,
        args: Option<Vec<ExprOrSpread>>,
    },
    Seq {
        exprs: Vec<Box<Expr>>,
    },
    Ident(Ident),
    Lit(Lit),
    Tpl(TplLit),
    Arrow {
        params: Vec<Pat>,
        body: BlockStmtOrExpr,
        is_generator: bool,
        is_async: bool,
    },
    Class(ClassExpr),
    Yield {
        arg: Option<Box<Expr>>,
        delegate: bool,
    },
    MetaProp {
        meta: Ident,
        prop: Ident,
    },
    Await {
        arg: Box<Expr>,
    },
    Paren(Box<Expr>),
}

Variants
ThisArray
Fields of Array

elems: Vec<Option<ExprOrSpread>>

Object
Fields of Object

props: Vec<Prop>

Function(FnExpr)Unary
Fields of Unary

op: UnaryOp

prefix: bool

arg: Box<Expr>

Update

++v, --v, v++, v--

Fields of Update

op: UpdateOp

prefix: bool

arg: Box<Expr>

Binary
Fields of Binary

op: BinaryOp

left: Box<Expr>

right: Box<Expr>

Assign
Fields of Assign

op: AssignOp

left: PatOrExpr

Pattern | Expr

right: Box<Expr>

Member

A member expression. If computed is true, the node corresponds to a computed (a[b]) member expression and property is an Expression. If computed is false, the node corresponds to a static (a.b) member expression and property is an Identifier.

Fields of Member

obj: ExprOrSuper

prop: Box<Expr>

computed: bool

Cond

true ? 'a' : 'b'

Fields of Cond

test: Box<Expr>

cons: Box<Expr>

alt: Box<Expr>

Call
Fields of Call

callee: ExprOrSuper

args: Vec<ExprOrSpread>

New

new Cat()

Fields of New

callee: Box<Expr>

args: Option<Vec<ExprOrSpread>>

None for new Cat.

Seq
Fields of Seq

exprs: Vec<Box<Expr>>

Ident(Ident)Lit(Lit)Tpl(TplLit)Arrow
Fields of Arrow

params: Vec<Pat>

body: BlockStmtOrExpr

is_generator: bool

is_async: bool

Class(ClassExpr)Yield
Fields of Yield

arg: Option<Box<Expr>>

delegate: bool

MetaProp
Fields of MetaProp

meta: Ident

prop: Ident

Await
Fields of Await

arg: Box<Expr>

Paren(Box<Expr>)