Enum swc_ecma_ast::Pat

pub enum Pat {
    Ident(BindingIdent),
    Array(ArrayPat),
    Rest(RestPat),
    Object(ObjectPat),
    Assign(AssignPat),
    Invalid(Invalid),
    Expr(Box<Expr>),
}

Variants

Ident(BindingIdent)

Array(ArrayPat)

Rest(RestPat)

Object(ObjectPat)

Assign(AssignPat)

Invalid(Invalid)

Expr(Box<Expr>)

Only for for-in / for-of loops. This is syntactically valid.

Implementations

impl Pat

pub fn is_ident(&self) -> bool

Returns true if self is of variant Ident.

pub fn as_ident(&self) -> Option<&BindingIdent>

Returns Some if self is a reference of variant Ident, and None otherwise.

pub fn as_mut_ident(&mut self) -> Option<&mut BindingIdent>

Returns Some if self is a mutable reference of variant Ident, and None otherwise.

pub fn expect_ident(self) -> BindingIdentwhere
Self: Debug,

Unwraps the value, yielding the content of Ident.

Panics

Panics if the value is not Ident, with a panic message including the content of self.

pub fn ident(self) -> Option<BindingIdent>

Returns Some if self is of variant Ident, and None otherwise.

pub fn is_array(&self) -> bool

Returns true if self is of variant Array.

pub fn as_array(&self) -> Option<&ArrayPat>

Returns Some if self is a reference of variant Array, and None otherwise.

pub fn as_mut_array(&mut self) -> Option<&mut ArrayPat>

Returns Some if self is a mutable reference of variant Array, and None otherwise.

pub fn expect_array(self) -> ArrayPatwhere
Self: Debug,

Unwraps the value, yielding the content of Array.

Panics

Panics if the value is not Array, with a panic message including the content of self.

pub fn array(self) -> Option<ArrayPat>

Returns Some if self is of variant Array, and None otherwise.

pub fn is_rest(&self) -> bool

Returns true if self is of variant Rest.

pub fn as_rest(&self) -> Option<&RestPat>

Returns Some if self is a reference of variant Rest, and None otherwise.

pub fn as_mut_rest(&mut self) -> Option<&mut RestPat>

Returns Some if self is a mutable reference of variant Rest, and None otherwise.

pub fn expect_rest(self) -> RestPatwhere
Self: Debug,

Unwraps the value, yielding the content of Rest.

Panics

Panics if the value is not Rest, with a panic message including the content of self.

pub fn rest(self) -> Option<RestPat>

Returns Some if self is of variant Rest, and None otherwise.

pub fn is_object(&self) -> bool

Returns true if self is of variant Object.

pub fn as_object(&self) -> Option<&ObjectPat>

Returns Some if self is a reference of variant Object, and None otherwise.

pub fn as_mut_object(&mut self) -> Option<&mut ObjectPat>

Returns Some if self is a mutable reference of variant Object, and None otherwise.

pub fn expect_object(self) -> ObjectPatwhere
Self: Debug,

Unwraps the value, yielding the content of Object.

Panics

Panics if the value is not Object, with a panic message including the content of self.

pub fn object(self) -> Option<ObjectPat>

Returns Some if self is of variant Object, and None otherwise.

pub fn is_assign(&self) -> bool

Returns true if self is of variant Assign.

pub fn as_assign(&self) -> Option<&AssignPat>

Returns Some if self is a reference of variant Assign, and None otherwise.

pub fn as_mut_assign(&mut self) -> Option<&mut AssignPat>

Returns Some if self is a mutable reference of variant Assign, and None otherwise.

pub fn expect_assign(self) -> AssignPatwhere
Self: Debug,

Unwraps the value, yielding the content of Assign.

Panics

Panics if the value is not Assign, with a panic message including the content of self.

pub fn assign(self) -> Option<AssignPat>

Returns Some if self is of variant Assign, and None otherwise.

pub fn is_invalid(&self) -> bool

Returns true if self is of variant Invalid.

pub fn as_invalid(&self) -> Option<&Invalid>

Returns Some if self is a reference of variant Invalid, and None otherwise.

pub fn as_mut_invalid(&mut self) -> Option<&mut Invalid>

Returns Some if self is a mutable reference of variant Invalid, and None otherwise.

pub fn expect_invalid(self) -> Invalidwhere
Self: Debug,

Unwraps the value, yielding the content of Invalid.

Panics

Panics if the value is not Invalid, with a panic message including the content of self.

pub fn invalid(self) -> Option<Invalid>

Returns Some if self is of variant Invalid, and None otherwise.

pub fn is_expr(&self) -> bool

Returns true if self is of variant Expr.

pub fn as_expr(&self) -> Option<&Box<Expr>>

Returns Some if self is a reference of variant Expr, and None otherwise.

pub fn as_mut_expr(&mut self) -> Option<&mut Box<Expr>>

Returns Some if self is a mutable reference of variant Expr, and None otherwise.

pub fn expect_expr(self) -> Box<Expr>where
Self: Debug,

Unwraps the value, yielding the content of Expr.

Panics

Panics if the value is not Expr, with a panic message including the content of self.

pub fn expr(self) -> Option<Box<Expr>>

Returns Some if self is of variant Expr, and None otherwise.