swc_ecma_ast

Enum Stmt

pub enum Stmt {

    Block(BlockStmt),
    Empty(EmptyStmt),
    Debugger(DebuggerStmt),
    With(WithStmt),
    Return(ReturnStmt),
    Labeled(LabeledStmt),
    Break(BreakStmt),
    Continue(ContinueStmt),
    If(IfStmt),
    Switch(SwitchStmt),
    Throw(ThrowStmt),
    Try(Box<TryStmt>),
    While(WhileStmt),
    DoWhile(DoWhileStmt),
    For(ForStmt),
    ForIn(ForInStmt),
    ForOf(ForOfStmt),
    Decl(Decl),
    Expr(ExprStmt),

}

Variants

Block(BlockStmt)

Empty(EmptyStmt)

Debugger(DebuggerStmt)

With(WithStmt)

Return(ReturnStmt)

Labeled(LabeledStmt)

Break(BreakStmt)

Continue(ContinueStmt)

If(IfStmt)

Switch(SwitchStmt)

Throw(ThrowStmt)

Try(Box<TryStmt>)

A try statement. If handler is null then finalizer must be a BlockStmt.

While(WhileStmt)

DoWhile(DoWhileStmt)

For(ForStmt)

ForIn(ForInStmt)

ForOf(ForOfStmt)

Decl(Decl)

Expr(ExprStmt)

Implementations

impl Stmt

pub const fn is_block(&self) -> bool

Returns true if self is of variant Block.

pub fn as_block(&self) -> Option<&BlockStmt>

Returns Some if self is a reference of variant Block, and None otherwise.

pub fn as_mut_block(&mut self) -> Option<&mut BlockStmt>

Returns Some if self is a mutable reference of variant Block, and None otherwise.

pub fn expect_block(self) -> BlockStmt
where Self: Debug,

Unwraps the value, yielding the content of Block.

Panics

Panics if the value is not Block, with a panic message including the content of self.

pub fn block(self) -> Option<BlockStmt>

Returns Some if self is of variant Block, and None otherwise.

pub const fn is_empty(&self) -> bool

Returns true if self is of variant Empty.

pub fn as_empty(&self) -> Option<&EmptyStmt>

Returns Some if self is a reference of variant Empty, and None otherwise.

pub fn as_mut_empty(&mut self) -> Option<&mut EmptyStmt>

Returns Some if self is a mutable reference of variant Empty, and None otherwise.

pub fn expect_empty(self) -> EmptyStmt
where Self: Debug,

Unwraps the value, yielding the content of Empty.

Panics

Panics if the value is not Empty, with a panic message including the content of self.

pub fn empty(self) -> Option<EmptyStmt>

Returns Some if self is of variant Empty, and None otherwise.

pub const fn is_debugger(&self) -> bool

Returns true if self is of variant Debugger.

pub fn as_debugger(&self) -> Option<&DebuggerStmt>

Returns Some if self is a reference of variant Debugger, and None otherwise.

pub fn as_mut_debugger(&mut self) -> Option<&mut DebuggerStmt>

Returns Some if self is a mutable reference of variant Debugger, and None otherwise.

pub fn expect_debugger(self) -> DebuggerStmt
where Self: Debug,

Unwraps the value, yielding the content of Debugger.

Panics

Panics if the value is not Debugger, with a panic message including the content of self.

pub fn debugger(self) -> Option<DebuggerStmt>

Returns Some if self is of variant Debugger, and None otherwise.

pub const fn is_with(&self) -> bool

Returns true if self is of variant With.

pub fn as_with(&self) -> Option<&WithStmt>

Returns Some if self is a reference of variant With, and None otherwise.

pub fn as_mut_with(&mut self) -> Option<&mut WithStmt>

Returns Some if self is a mutable reference of variant With, and None otherwise.

pub fn expect_with(self) -> WithStmt
where Self: Debug,

Unwraps the value, yielding the content of With.

Panics

Panics if the value is not With, with a panic message including the content of self.

pub fn with(self) -> Option<WithStmt>

Returns Some if self is of variant With, and None otherwise.

pub const fn is_return_stmt(&self) -> bool

Returns true if self is of variant Return.

pub fn as_return_stmt(&self) -> Option<&ReturnStmt>

Returns Some if self is a reference of variant Return, and None otherwise.

pub fn as_mut_return_stmt(&mut self) -> Option<&mut ReturnStmt>

Returns Some if self is a mutable reference of variant Return, and None otherwise.

pub fn expect_return_stmt(self) -> ReturnStmt
where Self: Debug,

Unwraps the value, yielding the content of Return.

Panics

Panics if the value is not Return, with a panic message including the content of self.

pub fn return_stmt(self) -> Option<ReturnStmt>

Returns Some if self is of variant Return, and None otherwise.

pub const fn is_labeled(&self) -> bool

Returns true if self is of variant Labeled.

pub fn as_labeled(&self) -> Option<&LabeledStmt>

Returns Some if self is a reference of variant Labeled, and None otherwise.

pub fn as_mut_labeled(&mut self) -> Option<&mut LabeledStmt>

Returns Some if self is a mutable reference of variant Labeled, and None otherwise.

pub fn expect_labeled(self) -> LabeledStmt
where Self: Debug,

Unwraps the value, yielding the content of Labeled.

Panics

Panics if the value is not Labeled, with a panic message including the content of self.

pub fn labeled(self) -> Option<LabeledStmt>

Returns Some if self is of variant Labeled, and None otherwise.

pub const fn is_break_stmt(&self) -> bool

Returns true if self is of variant Break.

pub fn as_break_stmt(&self) -> Option<&BreakStmt>

Returns Some if self is a reference of variant Break, and None otherwise.

pub fn as_mut_break_stmt(&mut self) -> Option<&mut BreakStmt>

Returns Some if self is a mutable reference of variant Break, and None otherwise.

pub fn expect_break_stmt(self) -> BreakStmt
where Self: Debug,

Unwraps the value, yielding the content of Break.

Panics

Panics if the value is not Break, with a panic message including the content of self.

pub fn break_stmt(self) -> Option<BreakStmt>

Returns Some if self is of variant Break, and None otherwise.

pub const fn is_continue_stmt(&self) -> bool

Returns true if self is of variant Continue.

pub fn as_continue_stmt(&self) -> Option<&ContinueStmt>

Returns Some if self is a reference of variant Continue, and None otherwise.

pub fn as_mut_continue_stmt(&mut self) -> Option<&mut ContinueStmt>

Returns Some if self is a mutable reference of variant Continue, and None otherwise.

pub fn expect_continue_stmt(self) -> ContinueStmt
where Self: Debug,

Unwraps the value, yielding the content of Continue.

Panics

Panics if the value is not Continue, with a panic message including the content of self.

pub fn continue_stmt(self) -> Option<ContinueStmt>

Returns Some if self is of variant Continue, and None otherwise.

pub const fn is_if_stmt(&self) -> bool

Returns true if self is of variant If.

pub fn as_if_stmt(&self) -> Option<&IfStmt>

Returns Some if self is a reference of variant If, and None otherwise.

pub fn as_mut_if_stmt(&mut self) -> Option<&mut IfStmt>

Returns Some if self is a mutable reference of variant If, and None otherwise.

pub fn expect_if_stmt(self) -> IfStmt
where Self: Debug,

Unwraps the value, yielding the content of If.

Panics

Panics if the value is not If, with a panic message including the content of self.

pub fn if_stmt(self) -> Option<IfStmt>

Returns Some if self is of variant If, and None otherwise.

pub const fn is_switch(&self) -> bool

Returns true if self is of variant Switch.

pub fn as_switch(&self) -> Option<&SwitchStmt>

Returns Some if self is a reference of variant Switch, and None otherwise.

pub fn as_mut_switch(&mut self) -> Option<&mut SwitchStmt>

Returns Some if self is a mutable reference of variant Switch, and None otherwise.

pub fn expect_switch(self) -> SwitchStmt
where Self: Debug,

Unwraps the value, yielding the content of Switch.

Panics

Panics if the value is not Switch, with a panic message including the content of self.

pub fn switch(self) -> Option<SwitchStmt>

Returns Some if self is of variant Switch, and None otherwise.

pub const fn is_throw(&self) -> bool

Returns true if self is of variant Throw.

pub fn as_throw(&self) -> Option<&ThrowStmt>

Returns Some if self is a reference of variant Throw, and None otherwise.

pub fn as_mut_throw(&mut self) -> Option<&mut ThrowStmt>

Returns Some if self is a mutable reference of variant Throw, and None otherwise.

pub fn expect_throw(self) -> ThrowStmt
where Self: Debug,

Unwraps the value, yielding the content of Throw.

Panics

Panics if the value is not Throw, with a panic message including the content of self.

pub fn throw(self) -> Option<ThrowStmt>

Returns Some if self is of variant Throw, and None otherwise.

pub const fn is_try_stmt(&self) -> bool

Returns true if self is of variant Try.

pub fn as_try_stmt(&self) -> Option<&Box<TryStmt>>

Returns Some if self is a reference of variant Try, and None otherwise.

pub fn as_mut_try_stmt(&mut self) -> Option<&mut Box<TryStmt>>

Returns Some if self is a mutable reference of variant Try, and None otherwise.

pub fn expect_try_stmt(self) -> Box<TryStmt>
where Self: Debug,

Unwraps the value, yielding the content of Try.

Panics

Panics if the value is not Try, with a panic message including the content of self.

pub fn try_stmt(self) -> Option<Box<TryStmt>>

Returns Some if self is of variant Try, and None otherwise.

pub const fn is_while_stmt(&self) -> bool

Returns true if self is of variant While.

pub fn as_while_stmt(&self) -> Option<&WhileStmt>

Returns Some if self is a reference of variant While, and None otherwise.

pub fn as_mut_while_stmt(&mut self) -> Option<&mut WhileStmt>

Returns Some if self is a mutable reference of variant While, and None otherwise.

pub fn expect_while_stmt(self) -> WhileStmt
where Self: Debug,

Unwraps the value, yielding the content of While.

Panics

Panics if the value is not While, with a panic message including the content of self.

pub fn while_stmt(self) -> Option<WhileStmt>

Returns Some if self is of variant While, and None otherwise.

pub const fn is_do_while(&self) -> bool

Returns true if self is of variant DoWhile.

pub fn as_do_while(&self) -> Option<&DoWhileStmt>

Returns Some if self is a reference of variant DoWhile, and None otherwise.

pub fn as_mut_do_while(&mut self) -> Option<&mut DoWhileStmt>

Returns Some if self is a mutable reference of variant DoWhile, and None otherwise.

pub fn expect_do_while(self) -> DoWhileStmt
where Self: Debug,

Unwraps the value, yielding the content of DoWhile.

Panics

Panics if the value is not DoWhile, with a panic message including the content of self.

pub fn do_while(self) -> Option<DoWhileStmt>

Returns Some if self is of variant DoWhile, and None otherwise.

pub const fn is_for_stmt(&self) -> bool

Returns true if self is of variant For.

pub fn as_for_stmt(&self) -> Option<&ForStmt>

Returns Some if self is a reference of variant For, and None otherwise.

pub fn as_mut_for_stmt(&mut self) -> Option<&mut ForStmt>

Returns Some if self is a mutable reference of variant For, and None otherwise.

pub fn expect_for_stmt(self) -> ForStmt
where Self: Debug,

Unwraps the value, yielding the content of For.

Panics

Panics if the value is not For, with a panic message including the content of self.

pub fn for_stmt(self) -> Option<ForStmt>

Returns Some if self is of variant For, and None otherwise.

pub const fn is_for_in(&self) -> bool

Returns true if self is of variant ForIn.

pub fn as_for_in(&self) -> Option<&ForInStmt>

Returns Some if self is a reference of variant ForIn, and None otherwise.

pub fn as_mut_for_in(&mut self) -> Option<&mut ForInStmt>

Returns Some if self is a mutable reference of variant ForIn, and None otherwise.

pub fn expect_for_in(self) -> ForInStmt
where Self: Debug,

Unwraps the value, yielding the content of ForIn.

Panics

Panics if the value is not ForIn, with a panic message including the content of self.

pub fn for_in(self) -> Option<ForInStmt>

Returns Some if self is of variant ForIn, and None otherwise.

pub const fn is_for_of(&self) -> bool

Returns true if self is of variant ForOf.

pub fn as_for_of(&self) -> Option<&ForOfStmt>

Returns Some if self is a reference of variant ForOf, and None otherwise.

pub fn as_mut_for_of(&mut self) -> Option<&mut ForOfStmt>

Returns Some if self is a mutable reference of variant ForOf, and None otherwise.

pub fn expect_for_of(self) -> ForOfStmt
where Self: Debug,

Unwraps the value, yielding the content of ForOf.

Panics

Panics if the value is not ForOf, with a panic message including the content of self.

pub fn for_of(self) -> Option<ForOfStmt>

Returns Some if self is of variant ForOf, and None otherwise.

pub const fn is_decl(&self) -> bool

Returns true if self is of variant Decl.

pub fn as_decl(&self) -> Option<&Decl>

Returns Some if self is a reference of variant Decl, and None otherwise.

pub fn as_mut_decl(&mut self) -> Option<&mut Decl>

Returns Some if self is a mutable reference of variant Decl, and None otherwise.

pub fn expect_decl(self) -> Decl
where Self: Debug,

Unwraps the value, yielding the content of Decl.

Panics

Panics if the value is not Decl, with a panic message including the content of self.

pub fn decl(self) -> Option<Decl>

Returns Some if self is of variant Decl, and None otherwise.

pub const fn is_expr(&self) -> bool

Returns true if self is of variant Expr.

pub fn as_expr(&self) -> Option<&ExprStmt>

Returns Some if self is a reference of variant Expr, and None otherwise.

pub fn as_mut_expr(&mut self) -> Option<&mut ExprStmt>

Returns Some if self is a mutable reference of variant Expr, and None otherwise.

pub fn expect_expr(self) -> ExprStmt
where Self: Debug,

Unwraps the value, yielding the content of Expr.

Panics

Panics if the value is not Expr, with a panic message including the content of self.

pub fn expr(self) -> Option<ExprStmt>

Returns Some if self is of variant Expr, and None otherwise.

impl Stmt

pub fn is_use_strict(&self) -> bool

pub fn can_precede_directive(&self) -> bool

Returns true if the statement does not prevent the directives below self from being directives.