Enum swc_ecma_ast::TsType

pub enum TsType {

    TsKeywordType(TsKeywordType),
    TsThisType(TsThisType),
    TsFnOrConstructorType(TsFnOrConstructorType),
    TsTypeRef(TsTypeRef),
    TsTypeQuery(TsTypeQuery),
    TsTypeLit(TsTypeLit),
    TsArrayType(TsArrayType),
    TsTupleType(TsTupleType),
    TsOptionalType(TsOptionalType),
    TsRestType(TsRestType),
    TsUnionOrIntersectionType(TsUnionOrIntersectionType),
    TsConditionalType(TsConditionalType),
    TsInferType(TsInferType),
    TsParenthesizedType(TsParenthesizedType),
    TsTypeOperator(TsTypeOperator),
    TsIndexedAccessType(TsIndexedAccessType),
    TsMappedType(TsMappedType),
    TsLitType(TsLitType),
    TsTypePredicate(TsTypePredicate),
    TsImportType(TsImportType),

}

Variants

TsKeywordType(TsKeywordType)

TsThisType(TsThisType)

TsFnOrConstructorType(TsFnOrConstructorType)

TsTypeRef(TsTypeRef)

TsTypeQuery(TsTypeQuery)

TsTypeLit(TsTypeLit)

TsArrayType(TsArrayType)

TsTupleType(TsTupleType)

TsOptionalType(TsOptionalType)

TsRestType(TsRestType)

TsUnionOrIntersectionType(TsUnionOrIntersectionType)

TsConditionalType(TsConditionalType)

TsInferType(TsInferType)

TsParenthesizedType(TsParenthesizedType)

TsTypeOperator(TsTypeOperator)

TsIndexedAccessType(TsIndexedAccessType)

TsMappedType(TsMappedType)

TsLitType(TsLitType)

TsTypePredicate(TsTypePredicate)

TsImportType(TsImportType)

Implementations

impl TsType

pub fn is_ts_keyword_type(&self) -> bool

Returns true if self is of variant TsKeywordType.

pub fn as_ts_keyword_type(&self) -> Option<&TsKeywordType>

Returns Some if self is a reference of variant TsKeywordType, and None otherwise.

pub fn as_mut_ts_keyword_type(&mut self) -> Option<&mut TsKeywordType>

Returns Some if self is a mutable reference of variant TsKeywordType, and None otherwise.

pub fn expect_ts_keyword_type(self) -> TsKeywordTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsKeywordType.

Panics

Panics if the value is not TsKeywordType, with a panic message including the content of self.

pub fn ts_keyword_type(self) -> Option<TsKeywordType>

Returns Some if self is of variant TsKeywordType, and None otherwise.

pub fn is_ts_this_type(&self) -> bool

Returns true if self is of variant TsThisType.

pub fn as_ts_this_type(&self) -> Option<&TsThisType>

Returns Some if self is a reference of variant TsThisType, and None otherwise.

pub fn as_mut_ts_this_type(&mut self) -> Option<&mut TsThisType>

Returns Some if self is a mutable reference of variant TsThisType, and None otherwise.

pub fn expect_ts_this_type(self) -> TsThisTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsThisType.

Panics

Panics if the value is not TsThisType, with a panic message including the content of self.

pub fn ts_this_type(self) -> Option<TsThisType>

Returns Some if self is of variant TsThisType, and None otherwise.

pub fn is_ts_fn_or_constructor_type(&self) -> bool

Returns true if self is of variant TsFnOrConstructorType.

pub fn as_ts_fn_or_constructor_type(&self) -> Option<&TsFnOrConstructorType>

Returns Some if self is a reference of variant TsFnOrConstructorType, and None otherwise.

pub fn as_mut_ts_fn_or_constructor_type(
&mut self
) -> Option<&mut TsFnOrConstructorType>

Returns Some if self is a mutable reference of variant TsFnOrConstructorType, and None otherwise.

pub fn expect_ts_fn_or_constructor_type(self) -> TsFnOrConstructorTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsFnOrConstructorType.

Panics

Panics if the value is not TsFnOrConstructorType, with a panic message including the content of self.

pub fn ts_fn_or_constructor_type(self) -> Option<TsFnOrConstructorType>

Returns Some if self is of variant TsFnOrConstructorType, and None otherwise.

pub fn is_ts_type_ref(&self) -> bool

Returns true if self is of variant TsTypeRef.

pub fn as_ts_type_ref(&self) -> Option<&TsTypeRef>

Returns Some if self is a reference of variant TsTypeRef, and None otherwise.

pub fn as_mut_ts_type_ref(&mut self) -> Option<&mut TsTypeRef>

Returns Some if self is a mutable reference of variant TsTypeRef, and None otherwise.

pub fn expect_ts_type_ref(self) -> TsTypeRefwhere
Self: Debug,

Unwraps the value, yielding the content of TsTypeRef.

Panics

Panics if the value is not TsTypeRef, with a panic message including the content of self.

pub fn ts_type_ref(self) -> Option<TsTypeRef>

Returns Some if self is of variant TsTypeRef, and None otherwise.

pub fn is_ts_type_query(&self) -> bool

Returns true if self is of variant TsTypeQuery.

pub fn as_ts_type_query(&self) -> Option<&TsTypeQuery>

Returns Some if self is a reference of variant TsTypeQuery, and None otherwise.

pub fn as_mut_ts_type_query(&mut self) -> Option<&mut TsTypeQuery>

Returns Some if self is a mutable reference of variant TsTypeQuery, and None otherwise.

pub fn expect_ts_type_query(self) -> TsTypeQuerywhere
Self: Debug,

Unwraps the value, yielding the content of TsTypeQuery.

Panics

Panics if the value is not TsTypeQuery, with a panic message including the content of self.

pub fn ts_type_query(self) -> Option<TsTypeQuery>

Returns Some if self is of variant TsTypeQuery, and None otherwise.

pub fn is_ts_type_lit(&self) -> bool

Returns true if self is of variant TsTypeLit.

pub fn as_ts_type_lit(&self) -> Option<&TsTypeLit>

Returns Some if self is a reference of variant TsTypeLit, and None otherwise.

pub fn as_mut_ts_type_lit(&mut self) -> Option<&mut TsTypeLit>

Returns Some if self is a mutable reference of variant TsTypeLit, and None otherwise.

pub fn expect_ts_type_lit(self) -> TsTypeLitwhere
Self: Debug,

Unwraps the value, yielding the content of TsTypeLit.

Panics

Panics if the value is not TsTypeLit, with a panic message including the content of self.

pub fn ts_type_lit(self) -> Option<TsTypeLit>

Returns Some if self is of variant TsTypeLit, and None otherwise.

pub fn is_ts_array_type(&self) -> bool

Returns true if self is of variant TsArrayType.

pub fn as_ts_array_type(&self) -> Option<&TsArrayType>

Returns Some if self is a reference of variant TsArrayType, and None otherwise.

pub fn as_mut_ts_array_type(&mut self) -> Option<&mut TsArrayType>

Returns Some if self is a mutable reference of variant TsArrayType, and None otherwise.

pub fn expect_ts_array_type(self) -> TsArrayTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsArrayType.

Panics

Panics if the value is not TsArrayType, with a panic message including the content of self.

pub fn ts_array_type(self) -> Option<TsArrayType>

Returns Some if self is of variant TsArrayType, and None otherwise.

pub fn is_ts_tuple_type(&self) -> bool

Returns true if self is of variant TsTupleType.

pub fn as_ts_tuple_type(&self) -> Option<&TsTupleType>

Returns Some if self is a reference of variant TsTupleType, and None otherwise.

pub fn as_mut_ts_tuple_type(&mut self) -> Option<&mut TsTupleType>

Returns Some if self is a mutable reference of variant TsTupleType, and None otherwise.

pub fn expect_ts_tuple_type(self) -> TsTupleTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsTupleType.

Panics

Panics if the value is not TsTupleType, with a panic message including the content of self.

pub fn ts_tuple_type(self) -> Option<TsTupleType>

Returns Some if self is of variant TsTupleType, and None otherwise.

pub fn is_ts_optional_type(&self) -> bool

Returns true if self is of variant TsOptionalType.

pub fn as_ts_optional_type(&self) -> Option<&TsOptionalType>

Returns Some if self is a reference of variant TsOptionalType, and None otherwise.

pub fn as_mut_ts_optional_type(&mut self) -> Option<&mut TsOptionalType>

Returns Some if self is a mutable reference of variant TsOptionalType, and None otherwise.

pub fn expect_ts_optional_type(self) -> TsOptionalTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsOptionalType.

Panics

Panics if the value is not TsOptionalType, with a panic message including the content of self.

pub fn ts_optional_type(self) -> Option<TsOptionalType>

Returns Some if self is of variant TsOptionalType, and None otherwise.

pub fn is_ts_rest_type(&self) -> bool

Returns true if self is of variant TsRestType.

pub fn as_ts_rest_type(&self) -> Option<&TsRestType>

Returns Some if self is a reference of variant TsRestType, and None otherwise.

pub fn as_mut_ts_rest_type(&mut self) -> Option<&mut TsRestType>

Returns Some if self is a mutable reference of variant TsRestType, and None otherwise.

pub fn expect_ts_rest_type(self) -> TsRestTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsRestType.

Panics

Panics if the value is not TsRestType, with a panic message including the content of self.

pub fn ts_rest_type(self) -> Option<TsRestType>

Returns Some if self is of variant TsRestType, and None otherwise.

pub fn is_ts_union_or_intersection_type(&self) -> bool

Returns true if self is of variant TsUnionOrIntersectionType.

pub fn as_ts_union_or_intersection_type(
&self
) -> Option<&TsUnionOrIntersectionType>

Returns Some if self is a reference of variant TsUnionOrIntersectionType, and None otherwise.

pub fn as_mut_ts_union_or_intersection_type(
&mut self
) -> Option<&mut TsUnionOrIntersectionType>

Returns Some if self is a mutable reference of variant TsUnionOrIntersectionType, and None otherwise.

pub fn expect_ts_union_or_intersection_type(self) -> TsUnionOrIntersectionTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsUnionOrIntersectionType.

Panics

Panics if the value is not TsUnionOrIntersectionType, with a panic message including the content of self.

pub fn ts_union_or_intersection_type(self) -> Option<TsUnionOrIntersectionType>

Returns Some if self is of variant TsUnionOrIntersectionType, and None otherwise.

pub fn is_ts_conditional_type(&self) -> bool

Returns true if self is of variant TsConditionalType.

pub fn as_ts_conditional_type(&self) -> Option<&TsConditionalType>

Returns Some if self is a reference of variant TsConditionalType, and None otherwise.

pub fn as_mut_ts_conditional_type(&mut self) -> Option<&mut TsConditionalType>

Returns Some if self is a mutable reference of variant TsConditionalType, and None otherwise.

pub fn expect_ts_conditional_type(self) -> TsConditionalTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsConditionalType.

Panics

Panics if the value is not TsConditionalType, with a panic message including the content of self.

pub fn ts_conditional_type(self) -> Option<TsConditionalType>

Returns Some if self is of variant TsConditionalType, and None otherwise.

pub fn is_ts_infer_type(&self) -> bool

Returns true if self is of variant TsInferType.

pub fn as_ts_infer_type(&self) -> Option<&TsInferType>

Returns Some if self is a reference of variant TsInferType, and None otherwise.

pub fn as_mut_ts_infer_type(&mut self) -> Option<&mut TsInferType>

Returns Some if self is a mutable reference of variant TsInferType, and None otherwise.

pub fn expect_ts_infer_type(self) -> TsInferTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsInferType.

Panics

Panics if the value is not TsInferType, with a panic message including the content of self.

pub fn ts_infer_type(self) -> Option<TsInferType>

Returns Some if self is of variant TsInferType, and None otherwise.

pub fn is_ts_parenthesized_type(&self) -> bool

Returns true if self is of variant TsParenthesizedType.

pub fn as_ts_parenthesized_type(&self) -> Option<&TsParenthesizedType>

Returns Some if self is a reference of variant TsParenthesizedType, and None otherwise.

pub fn as_mut_ts_parenthesized_type(
&mut self
) -> Option<&mut TsParenthesizedType>

Returns Some if self is a mutable reference of variant TsParenthesizedType, and None otherwise.

pub fn expect_ts_parenthesized_type(self) -> TsParenthesizedTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsParenthesizedType.

Panics

Panics if the value is not TsParenthesizedType, with a panic message including the content of self.

pub fn ts_parenthesized_type(self) -> Option<TsParenthesizedType>

Returns Some if self is of variant TsParenthesizedType, and None otherwise.

pub fn is_ts_type_operator(&self) -> bool

Returns true if self is of variant TsTypeOperator.

pub fn as_ts_type_operator(&self) -> Option<&TsTypeOperator>

Returns Some if self is a reference of variant TsTypeOperator, and None otherwise.

pub fn as_mut_ts_type_operator(&mut self) -> Option<&mut TsTypeOperator>

Returns Some if self is a mutable reference of variant TsTypeOperator, and None otherwise.

pub fn expect_ts_type_operator(self) -> TsTypeOperatorwhere
Self: Debug,

Unwraps the value, yielding the content of TsTypeOperator.

Panics

Panics if the value is not TsTypeOperator, with a panic message including the content of self.

pub fn ts_type_operator(self) -> Option<TsTypeOperator>

Returns Some if self is of variant TsTypeOperator, and None otherwise.

pub fn is_ts_indexed_access_type(&self) -> bool

Returns true if self is of variant TsIndexedAccessType.

pub fn as_ts_indexed_access_type(&self) -> Option<&TsIndexedAccessType>

Returns Some if self is a reference of variant TsIndexedAccessType, and None otherwise.

pub fn as_mut_ts_indexed_access_type(
&mut self
) -> Option<&mut TsIndexedAccessType>

Returns Some if self is a mutable reference of variant TsIndexedAccessType, and None otherwise.

pub fn expect_ts_indexed_access_type(self) -> TsIndexedAccessTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsIndexedAccessType.

Panics

Panics if the value is not TsIndexedAccessType, with a panic message including the content of self.

pub fn ts_indexed_access_type(self) -> Option<TsIndexedAccessType>

Returns Some if self is of variant TsIndexedAccessType, and None otherwise.

pub fn is_ts_mapped_type(&self) -> bool

Returns true if self is of variant TsMappedType.

pub fn as_ts_mapped_type(&self) -> Option<&TsMappedType>

Returns Some if self is a reference of variant TsMappedType, and None otherwise.

pub fn as_mut_ts_mapped_type(&mut self) -> Option<&mut TsMappedType>

Returns Some if self is a mutable reference of variant TsMappedType, and None otherwise.

pub fn expect_ts_mapped_type(self) -> TsMappedTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsMappedType.

Panics

Panics if the value is not TsMappedType, with a panic message including the content of self.

pub fn ts_mapped_type(self) -> Option<TsMappedType>

Returns Some if self is of variant TsMappedType, and None otherwise.

pub fn is_ts_lit_type(&self) -> bool

Returns true if self is of variant TsLitType.

pub fn as_ts_lit_type(&self) -> Option<&TsLitType>

Returns Some if self is a reference of variant TsLitType, and None otherwise.

pub fn as_mut_ts_lit_type(&mut self) -> Option<&mut TsLitType>

Returns Some if self is a mutable reference of variant TsLitType, and None otherwise.

pub fn expect_ts_lit_type(self) -> TsLitTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsLitType.

Panics

Panics if the value is not TsLitType, with a panic message including the content of self.

pub fn ts_lit_type(self) -> Option<TsLitType>

Returns Some if self is of variant TsLitType, and None otherwise.

pub fn is_ts_type_predicate(&self) -> bool

Returns true if self is of variant TsTypePredicate.

pub fn as_ts_type_predicate(&self) -> Option<&TsTypePredicate>

Returns Some if self is a reference of variant TsTypePredicate, and None otherwise.

pub fn as_mut_ts_type_predicate(&mut self) -> Option<&mut TsTypePredicate>

Returns Some if self is a mutable reference of variant TsTypePredicate, and None otherwise.

pub fn expect_ts_type_predicate(self) -> TsTypePredicatewhere
Self: Debug,

Unwraps the value, yielding the content of TsTypePredicate.

Panics

Panics if the value is not TsTypePredicate, with a panic message including the content of self.

pub fn ts_type_predicate(self) -> Option<TsTypePredicate>

Returns Some if self is of variant TsTypePredicate, and None otherwise.

pub fn is_ts_import_type(&self) -> bool

Returns true if self is of variant TsImportType.

pub fn as_ts_import_type(&self) -> Option<&TsImportType>

Returns Some if self is a reference of variant TsImportType, and None otherwise.

pub fn as_mut_ts_import_type(&mut self) -> Option<&mut TsImportType>

Returns Some if self is a mutable reference of variant TsImportType, and None otherwise.

pub fn expect_ts_import_type(self) -> TsImportTypewhere
Self: Debug,

Unwraps the value, yielding the content of TsImportType.

Panics

Panics if the value is not TsImportType, with a panic message including the content of self.

pub fn ts_import_type(self) -> Option<TsImportType>

Returns Some if self is of variant TsImportType, and None otherwise.