swc_ecma_ast

Struct Ident

pub struct Ident {
    pub span: Span,
    pub ctxt: SyntaxContext,
    pub sym: Atom,
    pub optional: bool,
}

A complete identifier with span.

Identifier of swc consists of two parts. The first one is symbol, which is stored using an interned string, Atom . The second one is SyntaxContext, which can be used to distinguish identifier with same symbol.

Let me explain this with an example.

let a = 5
{
    let a = 3;
}

In the code above, there are two variables with the symbol a.

Other compilers typically uses type like Scope, and store them nested, but in rust, type like Scope requires [Arc<Mutex>] so swc uses different approach. Instead of passing scopes, swc annotates two variables with different tag, which is named SyntaxContext. The notation for the syntax context is #n where n is a number. e.g. foo#1

For the example above, after applying resolver pass, it becomes.

let a#1 = 5
{
    let a#2 = 3;
}

Thanks to the tag we attached, we can now distinguish them.

(Atom, SyntaxContext)

See Id, which is a type alias for this.

This can be used to store all variables in a module to single hash map.

Comparison

While comparing two identifiers, you can use .to_id().

HashMap

There’s a type named Id which only contains minimal information to distinguish identifiers.

Fields
span: Spanctxt: SyntaxContextsym: Atomoptional: bool

TypeScript only. Used in case of an optional parameter.

Implementations

impl Ident

pub fn within_ignored_ctxt<F, Ret>(op: F) -> Ret
where F: FnOnce() -> Ret,

In op, EqIgnoreSpan of Ident will ignore the syntax context.

pub fn without_loc(self) -> Ident

Preserve syntax context while drop span.lo and span.hi.

pub fn to_id(&self) -> Id

Creates Id using Atom and SyntaxContext of self.

pub fn is_valid_ascii_start(c: u8) -> bool

pub fn is_valid_non_ascii_start(c: char) -> bool

pub fn is_valid_start(c: char) -> bool

Returns true if c is a valid character for an identifier start.

pub fn is_valid_non_ascii_continue(c: char) -> bool

pub fn is_valid_ascii_continue(c: u8) -> bool

pub fn is_valid_continue(c: char) -> bool

Returns true if c is a valid character for an identifier part after start.

pub fn verify_symbol(s: &str) -> Result<(), String>

Alternative for toIdentifier of babel.

Returns Ok if it’s a valid identifier and Err if it’s not valid. The returned Err contains the valid symbol.

pub fn with_prefix(&self, prefix: &str) -> Ident

Create a new identifier with the given prefix.

pub fn into_private(self) -> Ident

Create a private identifier that is unique in the file, but with the same symbol.

pub fn is_dummy(&self) -> bool

pub fn with_pos(self, lo: BytePos, hi: BytePos) -> Ident

Create a new identifier with the given position.

impl Ident

pub const fn new(sym: Atom, span: Span, ctxt: SyntaxContext) -> Self

pub fn new_private(sym: Atom, span: Span) -> Self

Creates a new private identifier. A private identifier is an identifier that is guaranteed to be unique.

See https://swc.rs/docs/contributing/es-commons/variable-management for more details.

Note: This method requires configuring GLOBALS because this method use Mark::new internally.

pub const fn new_no_ctxt(sym: Atom, span: Span) -> Self